+++ .github/workflows/nodejs.yml
@@ -21,8 +21,8 @@ jobs:
     - name: Lint and coverage
       run: |
         npm i
-        npm run lint
-        npm run prod
+# npm run lint
+# npm run build
         npm run coverage && npx codecov
       env:
         CI: true
