+++ .npmignore
@@ -4,19 +4,18 @@
 # Node modules
 node_modules/
 
+# Sources & Build
+src/
+tsconfig.json
+
 # Test & Coverage
 coverage/
 test/
+.eslintrc.json
+babel.config.js
 jest.config.js
 
 # Git
 .git/
 .gitignore
-
-# Other
-.eslintrc.js
 .github/
-
-
-# Webview
-docs/
