+++ calendars/ru.d.ts
@@ -1,0 +1,4 @@
+declare const CalendarSource: { [key: string]: Array<string> }
+
+export default CalendarSource
+
