+++ index.js
@@ -1,1 +1,37 @@
-(()=>{var r={831:r=>{r.exports={1999:[["1","2","3","4","6*","7","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","30*"],["1","2","3","4","8","9","10","15","16","22","23","29","30"],["5","6","11*","12","13","14","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["6","7","8","13","14","20","21","27","28"],["4","5","11","12","13","18","19","25","26","31*"]],2e3:[["1","2","3","4","6*","7","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["4","5","7*","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["1","2","6","7","8*","9","13","14","20","21","27","28"],["3","4","10","11","12","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["4","5","7","11","12","18","19","25","26"],["2","3","9","10","11*","12","16","17","23","24","30","31"]],2001:[["1","2","6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25"],["3","4","7*","8","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29","30*"],["1","2","5","6","8*","9","12","13","19","20","26","27"],["2","3","9","10","11*","12","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","6*","7","10","11","17","18","24","25"],["1","2","8","9","12","15","16","22","23","29","30"]],2002:[["1","2","5","6","7","12","13","19","20","26","27"],["2","3","9","10","16","17","22*","23","24","25"],["2","3","7*","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","28","30*"],["1","2","3","4","5","8*","9","10","11","12","19","25","26"],["1","2","8","9","11*","12","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","6*","7","8","9","16","17","23","24","30"],["1","7","8","11*","12","13","14","21","22","28","29","31*"]],2003:[["1","2","3","5*","6","7","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","24"],["1","2","7*","8","9","10","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27","30*"],["1","2","3","4","8*","9","10","11","17","18","24","25","31"],["1","7","8","11*","12","13","14","15","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26"],["1","2","6*","7","8","9","15","16","22","23","29","30"],["6","7","11*","12","13","14","20","21","27","28","31*"]],2004:[["1","2","3","4","6*","7","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","23","28","29"],["6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","30*"],["1","2","3","4","8","9","10","15","16","22","23","29","30"],["5","6","11*","12","13","14","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["6","7","8","13","14","20","21","27","28"],["4","5","11","12","13","18","19","25","26","31*"]],2005:[["1","2","3","4","5","6","7","8","9","10","15","16","22","23","29","30"],["5","6","12","13","19","20","22*","23","26","27"],["5*","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","2","7","8","9","14","15","21","22","28","29"],["4","5","11","12","13","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["3*","4","5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"]],2006:[["1","2","3","4","5","6","7","8","9","14","15","21","22","28","29"],["4","5","11","12","18","19","22*","23","24","25"],["4","5","7*","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["1","6*","7","8","9","13","14","20","21","27","28"],["3","4","10","11","12","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["3*","4","5","6","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"]],2007:[["1","2","3","4","5","6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","22*","23","24","25"],["3","4","7*","8","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28*","29","30"],["1","5","6","8*","9","12","13","19","20","26","27"],["2","3","9*","10","11","12","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","5","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29*","30","31"]],2008:[["1","2","3","4","5","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","22*","23","24","25"],["1","2","7*","8","9","10","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27","30*"],["1","2","3","8*","9","10","11","17","18","24","25","31"],["1","8","11*","12","13","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26"],["1*","2","3","4","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28","31*"]],2009:[["1","2","3","4","5","6","7","8","9","10","17","18","24","25","31"],["1","7","8","14","15","21","22","23","28"],["1","7","8","9","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["1","2","3","8*","9","10","11","16","17","23","24","30","31"],["6","7","11*","12","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","3*","4","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27","31*"]],2010:[["1","2","3","4","5","6","7","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","22","23","27*","28"],["6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","30*"],["1","2","3","8","9","10","15","16","22","23","29","30"],["5","6","11*","12","13","14","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["3*","4","5","6","7","14","20","21","27","28"],["4","5","11","12","18","19","25","26","31*"]],2011:[["1","2","3","4","5","6","7","8","9","10","15","16","22","23","29","30"],["5","6","12","13","19","20","22*","23","26","27"],["5*","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","2","7","8","9","14","15","21","22","28","29"],["4","5","11","12","13","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["3*","4","5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"]],2012:[["1","2","3","4","5","6","7","8","9","14","15","21","22","28","29"],["4","5","11","12","18","19","22*","23","25","26"],["3","4","7*","8","9","10","17","18","24","25","31"],["1","7","8","14","15","21","22","28*","29","30"],["1","6","7","8","9","12*","13","19","20","26","27"],["2","3","9*","10","11","12","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","5","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29*","30","31"]],2013:[["1","2","3","4","5","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","22*","23","24"],["2","3","7*","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["1","2","3","4","5","8*","9","10","11","12","18","19","25","26"],["1","2","8","9","11*","12","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","4","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29","31*"]],2014:[["1","2","3","4","5","6","7","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","24*"],["1","2","7*","8","9","10","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27","30*"],["1","2","3","4","8*","9","10","11","17","18","24","25","31"],["1","7","8","11*","12","13","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26"],["1","2","3","4","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"]],2015:[["1","2","3","4","5","6","7","8","9","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","23","28"],["1","7","8","9","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26","30*"],["1","2","3","4","8*","9","10","11","16","17","23","24","30","31"],["6","7","11*","12","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","3*","4","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27","31*"]],2016:[["1","2","3","4","5","6","7","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20*","21","22","23","27","28"],["5","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","2","3","7","8","9","14","15","21","22","28","29"],["4","5","11","12","13","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["3*","4","5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"]],2017:[["1","2","3","4","5","6","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","22*","23","24","25","26"],["4","5","7*","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["1","6","7","8","9","13","14","20","21","27","28"],["3","4","10","11","12","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["3*","4","5","6","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"]],2018:[["1","2","3","4","5","6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","22*","23","24","25"],["3","4","7*","8","9","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28*","29","30"],["1","2","5","6","8*","9","12","13","19","20","26","27"],["2","3","9*","10","11","12","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","5","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29*","30","31"]],2019:[["1","2","3","4","5","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","22*","23","24"],["2","3","7*","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28","30*"],["1","2","3","4","5","8*","9","10","11","12","18","19","25","26"],["1","2","8","9","11*","12","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","4","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29","31*"]],2020:[["1","2","3","4","5","6","7","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","24","29"],["1","7","8","9","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26","30*"],["1","2","3","4","5","8*","9","10","11","16","17","23","24","30","31"],["6","7","11*","12","13","14","20","21","24","27","28"],["1","4","5","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","3*","4","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27","31*"]],2021:[["1","2","3","4","5","6","7","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20*","21","22","23","27","28"],["6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","30*"],["1","2","3","4","5","6","7","8","9","10","15","16","22","23","29","30"],["5","6","11*","12","13","14","19","20","26","27"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["3*","4","5","6","7","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26","30*","31"]],2022:[["1","2","3","4","5","6","7","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","22*","23","26","27"],["5*","6","7","8","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","2","3","7","8","9","10","14","15","21","22","28","29"],["4","5","11","12","13","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["3*","4","5","6","12","13","19","20","26","27"],["3","4","10","11","17","18","24","25","31"]],2023:[["1","2","3","4","5","6","7","8","14","15","21","22","28","29"],["4","5","11","12","18","19","22*","23","25","26"],["4","5","7*","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","29","30"],["1","6","7","8*","9","13","14","20","21","27","28"],["3","4","10","11","12","17","18","24","25"],["1","2","8","9","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29"],["3*","4","5","6","11","12","18","19","25","26"],["2","3","9","10","16","17","23","24","30","31"]],2024:[["1","2","3","4","5","6","7","8","13","14","20","21","27","28"],["3","4","10","11","17","18","22*","23","24","25"],["2","3","7*","8","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28","30*"],["1","4","5","8*","9","11","12","18","19","25","26"],["1","2","8","9","11*","12","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28"],["3","4","10","11","17","18","24","25","31"],["1","7","8","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","4","9","10","16","17","23","24","30"],["1","7","8","14","15","21","22","28","29","31*"]],2025:[["1","2","3","4","5","6","7","8","11","12","18","19","25","26"],["1","2","8","9","15","16","22","23","24"],["1","2","7*","8","9","10","15","16","22","23","29","30"],["5","6","12","13","19","20","26","27","30*"],["1","3","4","8*","9","10","11","17","18","24","25","31"],["1","7","8","11*","12","14","15","21","22","28","29"],["5","6","12","13","19","20","26","27"],["2","3","9","10","16","17","23","24","30","31"],["6","7","13","14","20","21","27","28"],["4","5","11","12","18","19","25","26"],["1","2","3*","4","8","9","15","16","22","23","29","30"],["6","7","13","14","20","21","27","28","31*"]]}}},e={};function t(o){var n=e[o];if(void 0!==n)return n.exports;var a=e[o]={exports:{}};return r[o](a,a.exports,t),a.exports}t.n=r=>{var e=r&&r.__esModule?()=>r.default:()=>r;return t.d(e,{a:e}),e},t.d=(r,e)=>{for(var o in e)t.o(e,o)&&!t.o(r,o)&&Object.defineProperty(r,o,{enumerable:!0,get:e[o]})},t.o=(r,e)=>Object.prototype.hasOwnProperty.call(r,e),(()=>{"use strict";var r=t(831),e=t.n(r);const o=(r,e=0)=>Array.from(Array(r).keys()).map((r=>r+e));class n{constructor(r){if(r&&"ru"!==r)throw new Error("Locales differ from RU is not supported yet")}getCalendar(r,t,a){if(!r)throw new Error("Year must be not null");if(r instanceof Date&&(a=r.getDate(),t=r.getMonth()+1,r=r.getFullYear()),t){if(a){let o=n.DAY_WORK;if(void 0!==e()[r.toString()]&&void 0!==e()[r.toString()][t-1]){const i=e()[r.toString()][t-1];o=i.includes(a.toString())?n.DAY_HOLIDAY:o,o=i.includes(a+"*")?n.DAY_WORK_REDUCED:o}return o}return o(new Date(r,t,0).getDate(),1).map((e=>this.getCalendar(r,t,e)))}return o(12,1).map((e=>this.getCalendar(r,e)))}}n.DAY_WORK="work",n.DAY_HOLIDAY="holiday",n.DAY_WORK_REDUCED="work_reduced"})()})();
+"use strict";
+const ru_1 = require("./calendars/ru");
+const getSequencedArray = (length, startFrom = 0) => Array.from(Array(length).keys()).map((i) => i + startFrom);
+class Calendar {
+    constructor(locale) {
+        if (locale && locale !== "ru") {
+            throw new Error("Locales differ from RU is not supported yet");
+        }
+    }
+    getCalendar(year, month, day) {
+        if (!year) {
+            throw new Error("Year must be not null");
+        }
+        if (year instanceof Date) {
+            day = year.getDate();
+            month = year.getMonth() + 1;
+            year = year.getFullYear();
+        }
+        if (month) {
+            if (day) {
+                let dayType = Calendar.DAY_WORK;
+                if (ru_1[year.toString()] !== undefined && ru_1[year.toString()][month - 1] !== undefined) {
+                    const monthArray = ru_1[year.toString()][month - 1];
+                    dayType = monthArray.includes(day.toString()) ? Calendar.DAY_HOLIDAY : dayType;
+                    dayType = monthArray.includes(day + "*") ? Calendar.DAY_WORK_REDUCED : dayType;
+                }
+                return dayType;
+            }
+            return getSequencedArray(new Date(year, month, 0).getDate(), 1).map(day => this.getCalendar(year, month, day));
+        }
+        return getSequencedArray(12, 1).map(month => this.getCalendar(year, month));
+    }
+}
+Calendar.DAY_WORK = "work";
+Calendar.DAY_HOLIDAY = "holiday";
+Calendar.DAY_WORK_REDUCED = "work_reduced";
+module.exports = Calendar;
