+++ package.json
@@ -1,13 +1,12 @@
 {
   "name": "prod-cal",
-  "version": "1.3.2",
+  "version": "1.3.3",
   "description": "Production Calendar",
   "main": "index.js",
   "scripts": {
     "lint": "npx eslint ./src ./test --fix",
-    "watch": "npx webpack --mode development --watch",
-    "dev": "npx webpack --mode development",
-    "prod": "npx webpack --mode production",
+    "watch": "tsc -p tsconfig.json --watch",
+    "build": "tsc -p tsconfig.json",
     "test": "npx jest",
     "coverage": "npx jest --collectCoverage"
   },
@@ -32,11 +31,6 @@
     "@typescript-eslint/eslint-plugin": "5.15.0",
     "eslint": "8.6.0",
     "jest": "^27.4.7",
-    "ts-loader": "9.2.8",
-    "webpack": "5.70.0",
-    "webpack-cli": "4.9.2"
-  },
-  "dependencies": {
     "typescript": "4.6.2"
   }
 }
